+++ src/fiuba/algo3/algocraft/DesgastePorValor.java
@@ -2,14 +2,14 @@ package fiuba.algo3.algocraft;
 
 public class DesgastePorValor extends Desgaste{
 
-    public DesgastePorValor(int durabilidadInicial) {
+    public DesgastePorValor(double durabilidadInicial) {
         this.durabilidad = durabilidadInicial;
     }
 
 
-    public void desgastar(int valor, double factor) {
+    public void desgastar(double valor, double factor) {
         if (durabilidad == 0) throw new DurabilidadCeroException();
-        durabilidad -= valor * factor;
+        durabilidad -=  valor * factor;
         if (durabilidad <= 0) durabilidad = 0;
     }
